dce-scf: Fix history pointer setting
Correct the setting of the history pointer in the compression and
decompressing files. The history pointer was over writing the low
byte of the pending output pointer thereby corrupting the pending
pointer and the history pointer. This was not observed in stateless
cases because neither history nor pending output needs to be saved in
those cases

JIRA issue#: QSDK-3130

Signed-off-by: Ahmed Mansour <ahmed.mansour@nxp.com>